리팩토링: PnpmStoreManager 서비스 추출
- Editor/Package/PnpmStoreManager.cs 신설 (공유 store 경로 + install 재시도 stage 정의)
- AITPackageBuilder.GetSharedPnpmStorePath는 1줄 위임으로 유지 (외부 호환)
- InstallStages는 IReadOnlyList로 노출해 요소 수정 방지
- AppsInToss.Editor.Package using 선언으로 Menu 서비스 스타일과 일관성 유지
