+++ Editor/Package/PnpmStoreManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f267db9dc31a47ab8740c6101e909256
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
